refactor(composer): move recipient parsing and naming beside the composer
The address parser, the contact lookup with its cache, and the recipient
search are pure functions that lived among 900 lines of template wiring. They
now sit next to the composer: `addresses.ts` for parsing, with a runnable
self-check, and `recipients.ts` for naming and search. Behaviour unchanged.
